Reorganize MyHelsinki Places importer tests
Refs AHTI-152
